PIAPROGRAMACION2/requirements.txt
PyPDF2==2.11.1
PyPDF2==3.0.1
python_nmap==0.7.1
shodan==1.28.0
